requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
Flask-RESTful==0.3.9
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
